[APB-4804][JR] Implement Basic Auth for terminations

+++ app/uk/gov/hmrc/agentclientauthorisation/controllers/AgencyInvitationsController.scala
@@ -273,8 +273,11 @@ class AgencyInvitationsController @Inject()(
     }
   }
 
-  def removeAllInvitationsAndReferenceForArn(arn: Arn): Action[AnyContent] =
-    onlyStride(appConfig.terminationStrideEnrolment) { implicit request =>
+  def removeAllInvitationsAndReferenceForArn(arn: Arn): Action[AnyContent] = Action.async { implicit request =>
+    val expectedAuth: BasicAuthentication = appConfig.expectedAuth
+    val basicAuth: Option[BasicAuthentication] = getBasicAuth(request.headers)
+
+    if (basicAuth.contains(expectedAuth)) {
       if (Arn.isValid(arn.value)) {
         (for {
           invitationsDeleted <- invitationsService.removeAllInvitationsForAgent(arn)
@@ -294,5 +297,6 @@ class AgencyInvitationsController @Inject()(
           }
         }
       } else Future successful genericBadRequest(s"Invalid Arn given by Stride user: ${arn.value}")
-    }
+    } else Future successful Unauthorized
+  }
 }
